Add `TRYEXPAND_KEEP_ARTIFACTS` env key for keeping artifacts upon test completion

+++ src/project.rs
@@ -113,7 +113,33 @@ impl Project {
 
 impl Drop for Project {
     fn drop(&mut self) {
-        // Remove artifacts from the run (on a best-effort basis):
-        let _ = fs::remove_dir_all(&self.dir);
+        let should_keep_artifacts = match should_keep_artifacts() {
+            Ok(should_keep_artifacts) => should_keep_artifacts,
+            Err(err) => {
+                eprintln!("warning: {err}");
+                false
+            }
+        };
+
+        if !should_keep_artifacts {
+            // Remove artifacts from the run (on a best-effort basis):
+            let _ = fs::remove_dir_all(&self.dir);
+        }
+    }
+}
+
+fn should_keep_artifacts() -> Result<bool> {
+    let key = crate::TRYEXPAND_KEEP_ARTIFACTS_ENV_KEY;
+    let Some(var) = std::env::var_os(key) else {
+        return Ok(false);
+    };
+    let value = var.to_string_lossy().to_lowercase().to_owned();
+    match value.as_str() {
+        "1" | "yes" | "true" => Ok(true),
+        "0" | "no" | "false" => Ok(false),
+        _ => Err(Error::UnrecognizedEnv {
+            key: key.to_owned(),
+            value,
+        }),
     }
 }
